telegram-agent/adk-agent/fallback_tools.py
#!/usr/bin/env python3
"""
Fallback Player Analysis Tools - Work without database

These tools provide basic player analysis functionality even when
the database is not available or psycopg2 is not installed.

Usage:
    import these as fallback tools when database tools fail
"""

def get_player_matchups_fallback(player_name: str) -> str:
    """
    Fallback version of get_player_matchups that works without database.
    """
    return f"""I'd be happy to analyze {player_name}'s matchups! 

However, I'm currently experiencing connectivity issues with the full tennis database. To provide you with the most accurate analysis, I'll need access to the complete prediction data.

In the meantime, here's what I can tell you about {player_name}:

• **Player Type**: Tennis player analysis
• **Request Type**: Matchup history and performance
• **Data Needed**: Historical match results and upcoming fixtures

**Next Steps:**
1. Try again in a few moments when database connectivity is restored
2. Check if the player name is spelled correctly
3. Try searching for the player using just their surname (e.g., "Djokovic" instead of "Novak Djokovic")

Once the database connection is restored, I'll be able to provide detailed matchup information including:
- Recent match history
- Head-to-head records  
- Performance statistics
- Upcoming fixtures
- Surface-specific performance"""

def analyze_player_performance_fallback(player_name: str) -> str:
    """
    Fallback version of analyze_player_performance that works without database.
    """
    return f"""I'd love to analyze {player_name}'s performance! 

Unfortunately, I'm having trouble accessing the performance analytics database right now. To give you comprehensive performance insights, I need the full historical data.

**What I Would Analyze:**
- Recent match results and win/loss ratios
- Performance trends over time
- Surface-specific statistics (hard, clay, grass)
- Tournament performance breakdown
- Consistency metrics and improvement areas

**To Get Complete Analysis:**
- Try your query again in a moment
- Ensure the player name is spelled correctly
- Try using just the surname (e.g., "Djokovic" vs "Novak Djokovic")

Once database connectivity is restored, I'll provide detailed performance analytics including win rates, form analysis, and statistical breakdowns for {player_name}."""

# Mock tools that can be used when database is unavailable
def create_mock_player_analysis_tools():
    """
    Create mock tool versions that provide helpful responses when database is unavailable.
    """
    
    def mock_get_player_matchups(*args, **kwargs):
        """Mock tool that returns helpful fallback message."""
        # Extract player name from arguments
        player_name = "Unknown Player"
        if args:
            player_name = args[0]
        elif 'player_name' in kwargs:
            player_name = kwargs['player_name']
        
        return get_player_matchups_fallback(player_name)
    
    def mock_analyze_player_performance(*args, **kwargs):
        """Mock tool that returns helpful fallback message."""
        # Extract player name from arguments
        player_name = "Unknown Player"
        if args:
            player_name = args[0]
        elif 'player_name' in kwargs:
            player_name = kwargs['player_name']
            
        return analyze_player_performance_fallback(player_name)
    
    return mock_get_player_matchups, mock_analyze_player_performance

if __name__ == "__main__":
    # Test the fallback functions
    print("🧪 Testing Fallback Player Analysis Tools")
    print("=" * 50)
    
    print("\n1. Testing get_player_matchups_fallback('Cirpanli'):")
    print("-" * 50)
    result1 = get_player_matchups_fallback("Cirpanli")
    print(result1)
    
    print("\n\n2. Testing analyze_player_performance_fallback('Cirpanli'):")
    print("-" * 50)  
    result2 = analyze_player_performance_fallback("Cirpanli")
    print(result2)
    
    print("\n\n3. Testing mock tools:")
    print("-" * 50)
    mock_get_player_matchups, mock_analyze_player_performance = create_mock_player_analysis_tools()
    
    print("Mock get_player_matchups:")
    print(mock_get_player_matchups("Test Player"))
    
    print("\nMock analyze_player_performance:")
    print(mock_analyze_player_performance("Test Player"))
