telegram-agent/adk-agent/requirements.txt
# Enhanced Tennis Prediction Agent Dependencies

# Core dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-telegram-bot>=20.0

# AI/ML dependencies
requests>=2.31.0

# MCP Server dependencies
mcp>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Note: The google-ai-generativelanguage and google-ai-generativegenerativelanguage 
# packages are included with google-generativeai but listed here for clarity
google-ai-generativelanguage>=0.4.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
